+++ tests/test_unit.py
@@ -4,6 +4,7 @@ import dataclasses
 import datetime
 import enum
 import json
+import re
 import typing
 import uuid
 from unittest import mock
@@ -101,6 +102,7 @@ def test_unit__dump__ok__nominal_case(types_object: Types) -> None:
         "nested": {"name": "bar"},
         "nesteds": [{"name": "hello"}, {"name": "world"}],
         "mapping": {"foo": "bar"},
+        "optional": None,
         "datetime_": "2018-11-01T14:23:43.123456",
     } == serializer.dump(types_object)
 
@@ -179,11 +181,14 @@ def test_unit__json_schema__ok__nominal_case() -> None:
     serializer = serpyco.Serializer(Types)
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Types",
         "definitions": {
             "test_unit.Simple": {
+                "comment": "test_unit.Simple",
                 "description": "Basic class.",
                 "properties": {"name": {"type": "string"}},
                 "required": ["name"],
+                "additionalProperties": True,
                 "type": "object",
             }
         },
@@ -229,6 +234,7 @@ def test_unit__json_schema__ok__nominal_case() -> None:
             "mapping",
             "datetime_",
         ],
+        "additionalProperties": True,
         "type": "object",
     } == serializer.json_schema()
 
@@ -239,13 +245,16 @@ def test_unit__json_schema__ok__with_many() -> None:
         "$schema": "http://json-schema.org/draft-04/schema#",
         "definitions": {
             "test_unit.Simple": {
+                "comment": "test_unit.Simple",
                 "description": "Basic class.",
                 "properties": {"name": {"type": "string"}},
                 "required": ["name"],
                 "type": "object",
+                "additionalProperties": True,
             }
         },
         "items": {
+            "comment": "test_unit.Types",
             "description": "Testing class for supported serializer types.",
             "properties": {
                 "boolean": {"type": "boolean"},
@@ -291,6 +300,7 @@ def test_unit__json_schema__ok__with_many() -> None:
                 "mapping",
                 "datetime_",
             ],
+            "additionalProperties": True,
             "type": "object",
         },
         "type": "array",
@@ -301,17 +311,21 @@ def test_unit__json_schema__ok__cycle() -> None:
     builder = serpyco.SchemaBuilder(First)
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.First",
         "definitions": {
             "test_unit.Second": {
+                "comment": "test_unit.Second",
                 "description": "Cycle test class",
                 "properties": {"first": {"$ref": "#"}},
                 "required": ["first"],
                 "type": "object",
+                "additionalProperties": True,
             }
         },
         "description": "Cycle test class",
         "properties": {"second": {"$ref": "#/definitions/test_unit.Second"}},
         "required": ["second"],
+        "additionalProperties": True,
         "type": "object",
     } == builder.json_schema()
     nested = builder.nested_builders()
@@ -355,11 +369,13 @@ def test_unit__union__ok__nominal_case() -> None:
 
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.WithUnion",
         "definitions": {},
         "description": "Union test class",
         "properties": {"foo": {"anyOf": [{"type": "string"}, {"type": "integer"}]}},
         "required": ["foo"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
     assert {"foo": 42} == serializer.dump(WithUnion(foo=42), validate=True)
@@ -381,6 +397,7 @@ def test_unit__tuple__ok__nominal_case() -> None:
     serializer = serpyco.Serializer(WithTuple)
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.WithTuple",
         "definitions": {},
         "description": "Tuple test class",
         "properties": {
@@ -392,6 +409,7 @@ def test_unit__tuple__ok__nominal_case() -> None:
             }
         },
         "required": ["tuple_"],
+        "additionalProperties": True,
         "type": "object",
     } == serializer.json_schema()
     assert WithTuple(tuple_=("foo", 1)) == serializer.load({"tuple_": ["foo", 1]})
@@ -407,11 +425,13 @@ def test_unit__uniform_tuple__ok__nominal_case() -> None:
     serializer = serpyco.Serializer(WithTuple)
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.WithTuple",
         "definitions": {},
         "description": "Tuple test class",
         "properties": {"tuple_": {"type": "array", "items": {"type": "string"}}},
         "required": ["tuple_"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
     assert WithTuple(tuple_=("foo", "bar")) == serializer.load(
         {"tuple_": ["foo", "bar"]}
@@ -453,19 +473,22 @@ def test_unit__string_field_format_and_validators__ok__nominal_case() -> None:
             min_length=3,
             max_length=24,
         )
-        nested: Nested
+        nested: Nested  # type:ignore
 
     serializer = serpyco.Serializer(WithStringField)
 
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.WithStringField",
         "description": "String field test class",
         "definitions": {
             "test_unit.Nested": {
+                "comment": "test_unit.Nested",
                 "description": "Nested",
                 "properties": {"name": {"type": "string", "format": "date-time"}},
                 "required": ["name"],
                 "type": "object",
+                "additionalProperties": True,
             }
         },
         "properties": {
@@ -480,6 +503,7 @@ def test_unit__string_field_format_and_validators__ok__nominal_case() -> None:
         },
         "required": ["foo", "nested"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
     assert serializer.load({"foo": "Foo@foo.bar", "nested": {"name": "foo"}})
@@ -498,11 +522,13 @@ def test_unit__number_field__ok__nominal_case() -> None:
 
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.WithNumberField",
         "definitions": {},
         "description": "Number field test class",
         "properties": {"foo": {"type": "integer", "minimum": 0, "maximum": 12}},
         "required": ["foo"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
     assert serializer.load({"foo": 5})
@@ -533,23 +559,61 @@ def test_unit__type_encoders__ok__nominal_case() -> None:
     assert {"name": "foo"} == serializer.dump(Simple(name="bar"))
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Simple",
         "description": "Basic class.",
         "definitions": {},
         "properties": {"name": {}},
         "required": ["name"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
     second = serpyco.Serializer(Simple)
     assert {"name": "bar"} == second.dump(Simple(name="bar"))
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Simple",
         "description": "Basic class.",
         "definitions": {},
         "properties": {"name": {"type": "string"}},
         "required": ["name"],
         "type": "object",
+        "additionalProperties": True,
     } == second.json_schema()
+
+    @dataclasses.dataclass
+    class Nest:
+        """Nest"""
+
+        name: str
+        nested: Simple = serpyco.nested_field(type_encoders={str: Encoder()})
+
+    serializer = serpyco.Serializer(Nest)
+    assert {"name": "bar", "nested": {"name": "foo"}} == serializer.dump(
+        Nest(name="bar", nested=Simple("bar"))
+    )
+    assert {
+        "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Nest",
+        "description": "Nest",
+        "definitions": {
+            "test_unit.Simple": {
+                "comment": "test_unit.Simple",
+                "description": "Basic class.",
+                "properties": {"name": {}},
+                "required": ["name"],
+                "type": "object",
+                "additionalProperties": True,
+            }
+        },
+        "properties": {
+            "name": {"type": "string"},
+            "nested": {"$ref": "#/definitions/test_unit.Simple"},
+        },
+        "required": ["name", "nested"],
+        "type": "object",
+        "additionalProperties": True,
+    } == serializer.json_schema()
 
 
 def test_unit__global_type_encoders__ok__nominal_case() -> None:
@@ -567,22 +631,26 @@ def test_unit__global_type_encoders__ok__nominal_case() -> None:
     assert {"name": "foo"} == serializer.dump(Simple(name="bar"))
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Simple",
         "description": "Basic class.",
         "definitions": {},
         "properties": {"name": {}},
         "required": ["name"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
     second = serpyco.Serializer(Simple)
     assert {"name": "foo"} == second.dump(Simple(name="bar"))
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Simple",
         "description": "Basic class.",
         "definitions": {},
         "properties": {"name": {}},
         "required": ["name"],
         "type": "object",
+        "additionalProperties": True,
     } == second.json_schema()
 
     serpyco.Serializer.unregister_global_type(str)
@@ -591,11 +659,13 @@ def test_unit__global_type_encoders__ok__nominal_case() -> None:
     assert {"name": "bar"} == third.dump(Simple(name="bar"))
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Simple",
         "description": "Basic class.",
         "definitions": {},
         "properties": {"name": {"type": "string"}},
         "required": ["name"],
         "type": "object",
+        "additionalProperties": True,
     } == third.json_schema()
 
 
@@ -610,10 +680,13 @@ def test_unit__ignore__ok__nominal_case():
     assert {} == serializer.dump(Ignore(foo="bar"))
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Ignore",
         "description": "Ignore test class",
         "definitions": {},
         "properties": {},
         "type": "object",
+        "additionalProperties": True,
+        "required": [],
     } == serializer.json_schema()
 
 
@@ -629,11 +702,13 @@ def test_unit__only__ok__nominal_case():
     assert {"foo": "bar"} == serializer.dump(Only(foo="bar", bar="foo"))
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Only",
         "description": "Only test class",
         "definitions": {},
         "properties": {"foo": {"type": "string"}},
         "required": ["foo"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
 
@@ -649,6 +724,7 @@ def test_unit__field_description_and_examples__ok__nominal_case():
     serializer = serpyco.Serializer(Desc)
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Desc",
         "definitions": {},
         "description": "Description test class",
         "properties": {
@@ -660,6 +736,7 @@ def test_unit__field_description_and_examples__ok__nominal_case():
         },
         "required": ["foo"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
 
@@ -675,6 +752,7 @@ def test_unit__field_default__ok__nominal_case():
     serializer = serpyco.Serializer(Desc)
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Desc",
         "definitions": {},
         "description": "Description test class",
         "properties": {
@@ -688,6 +766,8 @@ def test_unit__field_default__ok__nominal_case():
             },
         },
         "type": "object",
+        "additionalProperties": True,
+        "required": [],
     } == serializer.json_schema()
 
 
@@ -741,11 +821,13 @@ def test_unit__exclude__ok__nominal_case():
     assert {"bar": "foo"} == serializer.dump(Exclude(foo="bar", bar="foo"))
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Exclude",
         "description": "Exclude test class",
         "definitions": {},
         "properties": {"bar": {"type": "string"}},
         "required": ["bar"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
 
@@ -770,16 +852,21 @@ def test_unit__nested_field__ok__nominal_case():
     )
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Parent",
         "description": "Parent test class",
         "definitions": {
             "test_unit.Nested_exclude_foo": {
                 "type": "object",
+                "additionalProperties": True,
+                "comment": "test_unit.Nested",
                 "description": "Nested test class",
                 "properties": {"bar": {"type": "string"}},
                 "required": ["bar"],
             },
             "test_unit.Nested_only_foo": {
                 "type": "object",
+                "additionalProperties": True,
+                "comment": "test_unit.Nested",
                 "description": "Nested test class",
                 "properties": {"foo": {"type": "string"}},
                 "required": ["foo"],
@@ -791,6 +878,7 @@ def test_unit__nested_field__ok__nominal_case():
         },
         "required": ["first", "second"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
 
@@ -846,7 +934,8 @@ def test_unit__rapidjson_validator__err_message():
         {"type": "object", "properties": {"name": {"type": "string"}}}
     )
     with pytest.raises(
-        serpyco.ValidationError, match=r'data\["name"\]: has type int, expected string'
+        serpyco.ValidationError,
+        match=r'value "42" at path "#/name" has type "int", expected "string"',
     ):
         val.validate({"name": 42})
 
@@ -888,18 +977,22 @@ def test_unit__custom_definition_name__ok__nominal_case():
 
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Class",
         "definitions": {
             "Custom": {
+                "comment": "test_unit.Nested",
                 "description": "Nested",
                 "properties": {"value": {"type": "integer"}},
                 "required": ["value"],
                 "type": "object",
+                "additionalProperties": True,
             }
         },
         "description": "Class",
         "properties": {"nested": {"$ref": "#/definitions/Custom"}},
         "required": ["nested"],
         "type": "object",
+        "additionalProperties": True,
     } == builder.json_schema()
     get_definition_name.assert_called_with(Nested, (), [], [])
 
@@ -936,11 +1029,15 @@ def test_unit__schema__ok__with_default_dataclass():
     serializer = serpyco.Serializer(Class)
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Class",
         "definitions": {
             "test_unit.Nested": {
+                "comment": "test_unit.Nested",
                 "description": "Nested",
                 "properties": {"name": {"default": "Hello", "type": "string"}},
                 "type": "object",
+                "additionalProperties": True,
+                "required": [],
             }
         },
         "description": "Class",
@@ -950,6 +1047,7 @@ def test_unit__schema__ok__with_default_dataclass():
         },
         "required": ["one"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
 
@@ -965,11 +1063,13 @@ def test_unit__generic_dataclass__ok__nominal_case():
     serializer = serpyco.Serializer(Gen[int])
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Gen",
         "definitions": {},
         "description": "Generic.",
         "properties": {"bar": {"type": "integer"}, "foo": {"type": "string"}},
         "required": ["foo", "bar"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
 
     @dataclasses.dataclass
@@ -979,19 +1079,23 @@ def test_unit__generic_dataclass__ok__nominal_case():
 
     serializer = serpyco.Serializer(WithGen)
     assert {
+        "comment": "test_unit.WithGen",
         "$schema": "http://json-schema.org/draft-04/schema#",
         "definitions": {
             "test_unit.Gen[int]": {
+                "comment": "test_unit.Gen",
                 "description": "Generic.",
                 "properties": {"bar": {"type": "integer"}, "foo": {"type": "string"}},
                 "required": ["foo", "bar"],
                 "type": "object",
+                "additionalProperties": True,
             }
         },
         "description": "With a generic.",
         "properties": {"nested": {"$ref": "#/definitions/test_unit.Gen[int]"}},
         "required": ["nested"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
     assert WithGen(nested=Gen(foo="bar", bar=12)) == serializer.load(
         {"nested": {"foo": "bar", "bar": 12}}
@@ -1011,6 +1115,7 @@ def test_unit__generic_dataclass__ok__nominal_case():
     serializer = serpyco.Serializer(SList[int])
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.SList",
         "definitions": {},
         "description": "List.",
         "properties": {
@@ -1019,6 +1124,7 @@ def test_unit__generic_dataclass__ok__nominal_case():
         },
         "required": ["items", "item_nb"],
         "type": "object",
+        "additionalProperties": True,
     } == serializer.json_schema()
     assert {"items": [0, 1], "item_nb": 2} == serializer.dump(SList([0, 1]))
 
@@ -1033,12 +1139,15 @@ def test_unit__schema__ok__none_default():
     serializer = serpyco.Serializer(Def)
     assert {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.Def",
         "definitions": {},
         "description": "Def.",
         "properties": {
             "foo": {"anyOf": [{"type": "integer"}, {"type": "null"}], "default": None}
         },
         "type": "object",
+        "required": [],
+        "additionalProperties": True,
     } == serializer.json_schema()
 
 
@@ -1083,10 +1192,12 @@ def test_unit__optional__custom_encoder__ok__nominal_case():
 
     assert serializer.json_schema() == {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.OptionalCustom",
         "definitions": {},
         "description": "OptionalCustom.",
         "properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}]}},
         "type": "object",
+        "additionalProperties": True,
         "required": ["name"],
     }
 
@@ -1118,11 +1229,13 @@ def test_unit__schema__ok__allowed_values():
 
     assert serializer.json_schema() == {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.WithAllowedValues",
         "definitions": {},
         "description": "WithAllowedValues.",
         "properties": {"foo": {"type": "integer", "enum": [1, 2, 3]}},
         "required": ["foo"],
         "type": "object",
+        "additionalProperties": True,
     }
 
 
@@ -1144,6 +1257,7 @@ def test_unit__schema__ok__allowed_values_with_enum():
 
     assert serializer.json_schema() == {
         "$schema": "http://json-schema.org/draft-04/schema#",
+        "comment": "test_unit.WithAllowedValues",
         "definitions": {},
         "description": "WithAllowedValues.",
         "properties": {
@@ -1151,6 +1265,7 @@ def test_unit__schema__ok__allowed_values_with_enum():
         },
         "required": ["foo"],
         "type": "object",
+        "additionalProperties": True,
     }
 
 
@@ -1206,15 +1321,18 @@ def test_unit__optional__ok__with_validator():
 def test_unit__optional__err__validation_error_message():
     @dataclasses.dataclass
     class Opt:
-        foo: typing.Optional[str] = serpyco.string_field(pattern="[A-Z]*")
+        foo: typing.Optional[str] = serpyco.string_field(min_length=5)
 
     serializer = serpyco.Serializer(Opt)
 
     with pytest.raises(
         serpyco.ValidationError,
-        match=r'data\["foo"\]: does not validate for any Union parameters.'
-        r" Details:\\n - string does not match pattern, "
-        r'got "bar", expected "\[A-Z\]\*"\\n - has type str, expected nul.',
+        match=re.escape(
+            r'value "bar" at path "#/foo" must match at least one of the '
+            r"following criteria:\n"
+            r"  - must have its length >= 5 but length is 3\n"
+            r"  - must be None"
+        ),
     ):
         serializer.load({"foo": "bar"})
 
@@ -1245,8 +1363,10 @@ def test_unit__validation__ok__several_errors():
     serializer = serpyco.Serializer(Foo)
     with pytest.raises(
         serpyco.ValidationError,
-        match=r'data\["bar"\]: has type int, expected string.\\n'
-        r'data\["foo"\]: has type str, expected integer.',
+        match=re.escape(
+            r'- value "12" at path "#/bar" has type "int", expected "string"\n'
+            r'- value "hello" at path "#/foo" has type "str", expected "integer"'
+        ),
     ):
         serializer.load({"bar": 12, "foo": "hello"})
 
@@ -1272,8 +1392,26 @@ def test_unit__validation_error_message__err__optional_sub_dataclass():
 
     with pytest.raises(
         serpyco.ValidationError,
-        match=r'data\["foo"\]: does not validate for any Union parameters. Details:\\n'
-        r' - is missing required properties "bar"\\n'
-        r" - has type dict, expected null.",
+        match=re.escape(
+            r'- value "{}" at path "#/foo" must match at least one of the '
+            r"following criteria:\n"
+            r'  - properties "bar" must be defined\n'
+            r"  - must be None"
+        ),
     ):
         serpyco.Serializer(Bar).load({"hello": 42, "foo": {}})
+
+
+def test_unit__strict_validation__err__additional_property():
+    @dataclasses.dataclass
+    class Foo:
+        bar: str
+
+    serializer = serpyco.Serializer(Foo, strict=True)
+    with pytest.raises(
+        serpyco.ValidationError,
+        match=re.escape(r'properties "hello" cannot be defined'),
+    ):
+        serializer.load({"hello": 42, "bar": "foo"})
+
+    assert serializer.load({"bar": "foo"}) == Foo("foo")